Framework: Add lockscreenwallpaper
Based on: http://gerrit.aokp.co/#/c/5133/

Unfortunately statusbar transparency does not work because
the lockscreen flag FLAG_SHOW_WALLPAPER is set, and when a custom
lockscreen wallpaper is set, the statusbar just shows the the user's
wallpaper. So, as a workaround, don't use FLAG_SHOW_WALLPAPER when a
custom wallpaper is being used.

Change-Id: I09cd8cd125daa2dbcd8e615a1afa5c0eb63b7ead